[FEAT] : 연락처 리스트뷰 추가 [CHORE] : fetchcontacts 함수 이동
